Funcionalidad para votar con AJAX

+++ app/Repositories/VoteRepository.php
@@ -17,6 +17,9 @@ class VoteRepository
 
     public function unvote(User $user, Ticket $ticket)
     {
+        if (! $user->hasVoted($ticket)) return false;
+
         $user->voted()->detach($ticket);
+        return true;
     }
 }
